pyproject.toml
[project]
name = "agent-skills-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "kaiser",email = "1587973663@qq.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "transformers (>=5.2.0,<6.0.0)",
    "openai (>=1.0.0,<2.0.0)",
    "pydantic (>=2.0.0,<3.0.0)",
    "fastapi (>=0.100.0,<1.0.0)",
    "uvicorn (>=0.20.0,<1.0.0)",
    "python-dotenv (>=1.0.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
